// apps/designer/src/types/database.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      categories: {
        Row: {
          account_id: string | null
          created_at: string
          default_og_image_url: string | null
          default_seo_description: string | null
          default_seo_title: string | null
          description: string | null
          id: string
          instance_type: string | null
          name: string
          status: string
          updated_at: string
          user_id: string | null
        }
        Insert: {
          account_id?: string | null
          created_at?: string
          default_og_image_url?: string | null
          default_seo_description?: string | null
          default_seo_title?: string | null
          description?: string | null
          id?: string
          instance_type?: string | null
          name: string
          status?: string
          updated_at?: string
          user_id?: string | null
        }
        Update: {
          account_id?: string | null
          created_at?: string
          default_og_image_url?: string | null
          default_seo_description?: string | null
          default_seo_title?: string | null
          description?: string | null
          id?: string
          instance_type?: string | null
          name?: string
          status?: string
          updated_at?: string
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "categories_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
        ]
      }
      categories_subcategories: {
        Row: {
          account_id: string | null
          category_id: string
          created_at: string
          credit_price: number
          demo_prompt_id: string | null
          demo_template_config: Json | null
          demo_theme_key: string | null
          description: string | null
          email_lead_price: number
          id: string
          instance_type: string | null
          phone_lead_price: number
          service_summary: string | null
          slug: string | null
          status: string
          subcategory: string
          updated_at: string
          use_cases: Json | null
          user_id: string | null
        }
        Insert: {
          account_id?: string | null
          category_id: string
          created_at?: string
          credit_price?: number
          demo_prompt_id?: string | null
          demo_template_config?: Json | null
          demo_theme_key?: string | null
          description?: string | null
          email_lead_price?: number
          id?: string
          instance_type?: string | null
          phone_lead_price?: number
          service_summary?: string | null
          slug?: string | null
          status?: string
          subcategory: string
          updated_at?: string
          use_cases?: Json | null
          user_id?: string | null
        }
        Update: {
          account_id?: string | null
          category_id?: string
          created_at?: string
          credit_price?: number
          demo_prompt_id?: string | null
          demo_template_config?: Json | null
          demo_theme_key?: string | null
          description?: string | null
          email_lead_price?: number
          id?: string
          instance_type?: string | null
          phone_lead_price?: number
          service_summary?: string | null
          slug?: string | null
          status?: string
          subcategory?: string
          updated_at?: string
          use_cases?: Json | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "categories_subcategories_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "categories_subcategories_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "categories_subcategories_demo_prompt_id_fkey"
            columns: ["demo_prompt_id"]
            isOneToOne: false
            referencedRelation: "prompts"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      user_accounts: {
        Row: {
          account_id: string
          created_at: string
          id: string
          status: string
          updated_at: string
          user_id: string
          user_status: string
        }
        Insert: {
          account_id: string
          created_at?: string
          id?: string
          status?: string
          updated_at?: string
          user_id: string
          user_status?: string
        }
        Update: {
          account_id?: string
          created_at?: string
          id?: string
          status?: string
          updated_at?: string
          user_id?: string
          user_status?: string
        }
        Relationships: [
          {
            foreignKeyName: "user_accounts_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}